Made the reactivation of jobs more robust.

+++ jobengine/main/management/commands/reactivate_jobs.py
@@ -21,24 +21,31 @@ class Command(BaseCommand):
         """
 
         all_crons = DBHelper.get_all_crons()
-        for job_entry in all_crons:
-            try:
-                job = Job(name=job_entry['name'], mode=job_entry['mode'], cron_string=job_entry['cron_string'],
-                          command=job_entry['command'], ipynb_file=job_entry['ipynb_file'])
-                if 'ipynb' in job.mode:
-                    print(f'[{timezone.now()}]{job.create_ipynb_cron()}')
-                else:
-                    print(f'[{timezone.now()}]{job.create_cron()}')
-            except Exception as e:
-                print(f'[{timezone.now()}] Server Error: {e}')
-        print(f'[{timezone.now()}] All jobs should be restarted.')
+        if len(all_crons) > 0:
+            for job_entry in all_crons:
+                try:
+                    job = Job(name=job_entry['name'], mode=job_entry['mode'], cron_string=job_entry['cron_string'],
+                              command=job_entry['command'], ipynb_file=job_entry['ipynb_file'])
+                    if 'ipynb' in job.mode:
+                        print(f'[{timezone.now()}]{job.create_ipynb_cron()}')
+                    else:
+                        print(f'[{timezone.now()}]{job.create_cron()}')
+                except Exception as e:
+                    print(f'[{timezone.now()}] Server Error: {e}')
+            print(f'[{timezone.now()}] All jobs should be restarted.')
+        else:
+            print(f'[{timezone.now()}] No jobs to restart.')
 
-        dir_list = os.listdir('/jobengine/jobs')
-        all_jobs = DBHelper.get_jobs()
-        for directory in dir_list:
-            if os.path.isdir(f'/jobengine/jobs/{directory}'):
-                if not any(entry['name'] == directory for entry in all_jobs):
-                    shutil.rmtree(f'/jobengine/jobs/{directory}')
-                    print(f'[{timezone.now()}] {directory} is not associated with a job and will be destroyed.')
-        print(f'[{timezone.now()}] job dir should be clean now.')
+        if not os.path.exists('/jobengine/jobs'):
+            os.makedirs('/jobengine/jobs')
+            print(f'[{timezone.now()}] Created dir for jobs.')
+        else:
+            dir_list = os.listdir('/jobengine/jobs')
+            all_jobs = DBHelper.get_jobs()
+            for directory in dir_list:
+                if os.path.isdir(f'/jobengine/jobs/{directory}'):
+                    if not any(entry['name'] == directory for entry in all_jobs):
+                        shutil.rmtree(f'/jobengine/jobs/{directory}')
+                        print(f'[{timezone.now()}] {directory} is not associated with a job and will be destroyed.')
+            print(f'[{timezone.now()}] job dir should be clean now.')
